Fixed a major error in foresight (See notes)
It occured to me that our classes would rely on the Robot class we had in the project and not the ones of the teams. Therefore, I changed how the classes interact with the robot, by making only an AbstractDrivetrain needed for implementation of the library. There, teams can use super(Robot.drivetrain) in their drivetrain's constructor to set the stones in motion.